+++ src/DAL/Entities/Assessment.cs
@@ -3,13 +3,16 @@ using System.Collections.Generic;
 
 namespace DAL.Entities;
 
-public partial class Assessment
+public partial class Assessment : DAL.Interfaces.IEntityScoped
 {
     public int Id { get; set; }
 
     public string Name { get; set; } = null!;
 
     public DateTime Created { get; set; }
+
+    public int? EntityId { get; set; }
+    public virtual Entity? Entity { get; set; }
 
     public virtual ICollection<AssessmentAnswer> AssessmentAnswers { get; set; } = new List<AssessmentAnswer>();