+++ src/DAL/Interfaces/IEntityScoped.cs
@@ -1,0 +1,6 @@
+namespace DAL.Interfaces;
+
+public interface IEntityScoped
+{
+    int? EntityId { get; set; }
+}